windows

+++ vedadet/ops/dcn/src/cuda/deform_conv_cuda_kernel.cu
@@ -86,8 +86,8 @@ __device__ scalar_t deformable_im2col_bilinear(const scalar_t *bottom_data, cons
                                                const int height, const int width, scalar_t h, scalar_t w)
 {
 
-  int h_low = floor(h);
-  int w_low = floor(w);
+  int h_low = floorf(h);
+  int w_low = floorf(w);
   int h_high = h_low + 1;
   int w_high = w_low + 1;
 
@@ -125,8 +125,8 @@ __device__ scalar_t get_gradient_weight(scalar_t argmax_h, scalar_t argmax_w,
     return 0;
   }
 
-  int argmax_h_low = floor(argmax_h);
-  int argmax_w_low = floor(argmax_w);
+  int argmax_h_low = floorf(argmax_h);
+  int argmax_w_low = floorf(argmax_w);
   int argmax_h_high = argmax_h_low + 1;
   int argmax_w_high = argmax_w_low + 1;
 
@@ -154,8 +154,8 @@ __device__ scalar_t get_coordinate_weight(scalar_t argmax_h, scalar_t argmax_w,
     return 0;
   }
 
-  int argmax_h_low = floor(argmax_h);
-  int argmax_w_low = floor(argmax_w);
+  int argmax_h_low = floorf(argmax_h);
+  int argmax_w_low = floorf(argmax_w);
   int argmax_h_high = argmax_h_low + 1;
   int argmax_w_high = argmax_w_low + 1;
 
@@ -468,8 +468,8 @@ template <typename scalar_t>
 __device__ scalar_t dmcn_im2col_bilinear(const scalar_t *bottom_data, const int data_width,
                                          const int height, const int width, scalar_t h, scalar_t w)
 {
-  int h_low = floor(h);
-  int w_low = floor(w);
+  int h_low = floorf(h);
+  int w_low = floorf(w);
   int h_high = h_low + 1;
   int w_high = w_low + 1;
 
@@ -506,8 +506,8 @@ __device__ scalar_t dmcn_get_gradient_weight(scalar_t argmax_h, scalar_t argmax_
     return 0;
   }
 
-  int argmax_h_low = floor(argmax_h);
-  int argmax_w_low = floor(argmax_w);
+  int argmax_h_low = floorf(argmax_h);
+  int argmax_w_low = floorf(argmax_w);
   int argmax_h_high = argmax_h_low + 1;
   int argmax_w_high = argmax_w_low + 1;
 
@@ -534,8 +534,8 @@ __device__ scalar_t dmcn_get_coordinate_weight(scalar_t argmax_h, scalar_t argma
     return 0;
   }
 
-  int argmax_h_low = floor(argmax_h);
-  int argmax_w_low = floor(argmax_w);
+  int argmax_h_low = floorf(argmax_h);
+  int argmax_w_low = floorf(argmax_w);
   int argmax_h_high = argmax_h_low + 1;
   int argmax_w_high = argmax_w_low + 1;
